+++ src/components/form/FSelect.vue
@@ -55,7 +55,7 @@ FField.FSelect(
               .FSelect__text {{ getValueLabel(fieldValue) }}
             .FSelect__placeholder(v-else) {{ placeholder }}
         FIcon.FSelect__icon(
-          v-if="!disabled && !loading"
+          v-if="!loading"
           :name="iconName"
           :size="16"
           color="transparent"
@@ -144,6 +144,9 @@ FField.FSelect(
   .FSelect__placeholder
     color var(--color--neutral--light-1)
 
+  .FSelect__icon
+    opacity 0.25
+
   &,
   &:hover,
   &:focus
@@ -161,7 +164,7 @@ FField.FSelect(
 
 .FSelect--small
   .FSelect__select
-    padding rem(12)
+    padding rem(12) rem(4) rem(12) rem(12)
     height rem(36)
     font-size rem(14)
 </style>
@@ -386,9 +389,16 @@ function getValueLabel(value: string): string | undefined {
 function handleIconClick(event: Event): void {
   if (props.clearable && isMenuOpen.value && fieldValue.value) {
     event.stopPropagation();
-    fieldValue.value = null;
-    emit('clear');
+    clear();
   }
+}
+
+/**
+ * Clear selection
+ */
+function clear(): void {
+  fieldValue.value = null;
+  emit('clear');
 }
 
 /**